backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Caching==2.3.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.4
requests==2.32.3
httpx==0.27.2
pydantic==2.9.2
openai==1.52.2
gunicorn==23.0.0
firebase-admin==6.5.0
google-cloud-storage==2.18.2
